+++ lib/squake/calculation.rb
@@ -10,13 +10,13 @@ module Squake
 
     sig do
       params(
-        client: Squake::Client,
         items: T::Array[T.any(Squake::Model::Items::BaseType, T::Hash[T.any(String, Symbol), T.untyped])],
         carbon_unit: String,
         expand: T::Array[String],
+        client: Squake::Client,
       ).returns(Squake::Return[Squake::Model::Carbon])
     end
-    def self.create(client:, items:, carbon_unit: 'gram', expand: [])
+    def self.create(items:, carbon_unit: 'gram', expand: [], client: Squake::Client.new)
       # @TODO: add typed objects for all possible items. Until then, we allow either a Hash or a T::Struct
       items = items.map do |item|
         item.is_a?(T::Struct) ? item.serialize : item