+++ lib/squake/calculation_with_pricing.rb
@@ -10,15 +10,15 @@ module Squake
 
     sig do
       params(
-        client: Squake::Client,
         items: T::Array[T.any(Squake::Model::Items::BaseType, T::Hash[T.any(String, Symbol), T.untyped])],
         product: String,
         currency: String,
         carbon_unit: String,
         expand: T::Array[String],
+        client: Squake::Client,
       ).returns(Squake::Return[Squake::Model::Pricing])
     end
-    def self.quote(client:, items:, product:, currency: 'EUR', carbon_unit: 'gram', expand: [])
+    def self.quote(items:, product:, currency: 'EUR', carbon_unit: 'gram', expand: [], client: Squake::Client.new)
       # @TODO: add typed objects for all possible items. Until then, we allow either a Hash or a T::Struct
       items = items.map do |item|
         item.is_a?(T::Struct) ? item.serialize : item